Rename pixels module to camera module
