+++ fact/instrument/camera.py
