+++ tests/test_all.py
@@ -38,7 +38,7 @@ def test_night_integer_pandas_datetimeindex():
 
 
 def test_patch_indices():
-    from fact.instrument.pixels import patch_indices
+    from fact.instrument.camera import patch_indices
 
     pi = patch_indices
     assert pi[pi.bias_patch_id==1].trigger_patch_id.iloc[0] == 0
@@ -51,7 +51,7 @@ def test_patch_indices():
 
 
 def test_easier_use_of_patch_indices():
-    from fact.instrument.pixels import patch_indices
+    from fact.instrument.camera import patch_indices
     pi = patch_indices
 
     bias_patch_ids = pi.bias_patch_id.values
@@ -73,7 +73,7 @@ def test_bias_patch_values_into_trigger_patches():
 
     # so you might start by sorting by the index you have and get the one you
     # want, like this:
-    from fact.instrument.pixels import patch_indices
+    from fact.instrument.camera import patch_indices
     pi = patch_indices
 
     t_id_BY_b_id = pi.sort_values('bias_patch_id').trigger_patch_id.values
@@ -95,8 +95,8 @@ def test_bias_patch_values_into_trigger_patches():
 
 def test_coords_relation_to_pos_from_dataframe():
 
-    from fact.instrument.pixels import get_pixel_dataframe
-    from fact.instrument.pixels import get_pixel_coords
+    from fact.instrument.camera import get_pixel_dataframe
+    from fact.instrument.camera import get_pixel_coords
     import numpy as np
 
     pc = get_pixel_coords()
